myblog_backup/requirements.txt
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==2.9.1
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
humanize==0.5.1
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
mysqlclient==1.3.10
netaddr==0.7.19
openvpn-status==0.2.0
passlib==1.7.1
PyMySQL==0.7.11
python-engineio==1.7.0
python-socketio==1.7.7
redis==2.10.6
requests==2.18.4
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
